+++ app/src/demo/table/columns.tsx
@@ -1,14 +1,11 @@
 import * as React from 'react';
 import { Text, Badge, EpamAdditionalColor, FlexRow, IconButton, LinkButton, Tag } from '@epam/promo';
-import { DataQueryFilter, DataColumnProps } from '@epam/uui';
-import { City, Department, Person, PersonGroup, Manager, Country, Office } from '@epam/uui-docs';
-import { ITableFilter, PersonTableRecordId } from './types';
+import { DataColumnProps } from "@epam/uui";
 import * as css from './DemoTable.scss';
 import { ReactComponent as ViewIcon } from '@epam/assets/icons/common/action-eye-18.svg';
-import { addFiltersToColumns } from "./helpers";
 
-export function getColumns(filters: ITableFilter[]) {
-    const personColumns: DataColumnProps<Person, PersonTableRecordId, DataQueryFilter<Person>>[] = [
+export function getColumns<TFilter extends Record<string, any>>(): DataColumnProps[] {
+    return [
         {
             key: 'name',
             caption: "Name",
@@ -22,7 +19,6 @@ export function getColumns(filters: ITableFilter[]) {
             caption: 'Profile Status',
             render: p => p.profileStatus && <FlexRow>
                 <Badge
-                    cx={ css.status }
                     fill="transparent"
                     color={ p.profileStatus.toLowerCase() as EpamAdditionalColor }
                     caption={ p.profileStatus }/>
@@ -37,9 +33,7 @@ export function getColumns(filters: ITableFilter[]) {
             key: 'jobTitle',
             caption: "Title",
             render: r => <Text>{ r.jobTitle }</Text>,
-            grow: 0,
-            shrink: 0,
-            width: 200,
+            minWidth: 200,
             isSortable: true,
             isFilterActive: f => !!f.jobTitleId,
         },
@@ -57,7 +51,7 @@ export function getColumns(filters: ITableFilter[]) {
         {
             key: 'officeAddress',
             caption: "Office",
-            render: p => <Text cx={ css.office }>{ p.officeAddress }</Text>,
+            render: p => <Text>{ p.officeAddress }</Text>,
             grow: 0,
             shrink: 0,
             width: 150,
@@ -142,18 +136,4 @@ export function getColumns(filters: ITableFilter[]) {
             fix: 'right',
         },
     ];
-
-    const groupColumns: DataColumnProps<PersonGroup, number, DataQueryFilter<Person>>[] = [
-        {
-            key: 'name',
-            caption: "Name",
-            render: p => <FlexRow><Text>{ p.name }</Text><Tag cx={ css.counter } count={ p.count }/></FlexRow>,
-            grow: 1,
-        },
-    ];
-
-    return {
-        personColumns: addFiltersToColumns(personColumns, filters),
-        groupColumns,
-    };
 }