+++ edu-core-routing/types.ts
@@ -1,4 +1,0 @@
-export interface Link {
-    pathname: string;
-    query?: any;
-}