+++ epam-promo/components/datePickers/index.ts
@@ -1,3 +1,5 @@
+export * from './Calendar';
+export * from './CalendarPresets';
 export * from './DatePicker';
 export * from './DatePickerBody';
 export * from './RangeDatePicker';