+++ loveship/components/navigation/MainMenu/MainMenuButton.tsx
@@ -1,7 +1,7 @@
 import * as css from './MainMenuButton.scss';
 import { Button, ButtonProps } from '@epam/uui-components';
 import { IAdaptiveItem, withMods } from '@epam/uui-core';
-import { ReactComponent as TriangleIcon } from '../../icons/triangle.svg';
+import { ReactComponent as TriangleIcon } from '../../icons/chevron-down-24.svg';
 
 export interface MainMenuButtonMods {
     type?: 'primary' | 'secondary';